Add cholesky decomp, Closes #30, and dynamic print buf

+++ matrix/Dcm.hpp
@@ -164,7 +164,7 @@ public:
     void renormalize()
     {
         /* renormalize rows */
-        for (int row = 0; row < 3; row++) {
+        for (size_t row = 0; row < 3; row++) {
             matrix::Vector3f rvec(this->_data[row]);
             this->setRow(row, rvec.normalized());
         }
