+++ test/setIdentity.cpp
@@ -10,8 +10,8 @@ int main()
     Matrix3f A;
     A.setIdentity();
 
-    for (int i = 0; i < 3; i++) {
-        for (int j = 0; j < 3; j++) {
+    for (size_t i = 0; i < 3; i++) {
+        for (size_t j = 0; j < 3; j++) {
             if (i == j) {
                 TEST( fabs(A(i, j) -  1) < 1e-7);
 
@@ -24,8 +24,8 @@ int main()
     Matrix3f B;
     B.identity();
 
-    for (int i = 0; i < 3; i++) {
-        for (int j = 0; j < 3; j++) {
+    for (size_t i = 0; i < 3; i++) {
+        for (size_t j = 0; j < 3; j++) {
             if (i == j) {
                 TEST( fabs(B(i, j) -  1) < 1e-7);
